began moving functions to separate files
